Update wait-for-postgres.sh
I made some modifications:

Added a timeout variable and a start_time variable to keep track of the elapsed time. We also added an if statement that checks if the elapsed time has exceeded the timeout value. If it has, the script exits with an error message.

Replaced the hardcoded database username and password with environment variables.

Added the set -o pipefail command to ensure that the script fails if any command in a pipeline fails. We also added some error messages to make it easier to diagnose problems.